gave every component its own folder
