+++ apps/web/src/components/Banner/Banner.scss
@@ -1,0 +1,46 @@
+.Banner {
+  @keyframes slideIn {
+    0% {
+      transform: translate(-50%, -100%);
+      opacity: 0;
+    }
+    100% {
+      transform: translate(-50%, 0);
+      opacity: 1;
+    }
+  }
+
+  position: fixed;
+  top: 15%;
+  left: 50%;
+  transform: translateX(-50%);
+  padding: 0.8em 2em;
+  border-radius: 8px;
+  font-family: "Poppins", "raleway", "rubik";
+  font-size: 1.1em;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.2);
+  animation: slideIn 0.3s ease-out forwards;
+  z-index: 1000;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.3em;
+  @media screen and (max-width: 460px) {
+    width: 50vw;
+  }
+  &.error {
+    background: linear-gradient(135deg, #e02b2b 20%, #ff4646 50%, #e02b2b);
+    color: #f9f9f9;
+  }
+
+  &.info {
+    background: linear-gradient(135deg, #ff77b4 20%, #9b4d97 50%, #ff77b4);
+    color: #f9f9f9;
+  }
+
+  span {
+    display: block;
+    text-align: center;
+    text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.2);
+  }
+}
