+++ apps/web/src/components/Navigation/NavMobile.scss
@@ -1,0 +1,62 @@
+$background: #f9f9f9;
+$pink: #ff77b4;
+$purple: #9b4d97;
+$dark-gray: #383838;
+$error-red: #e02b2b;
+$border-radius: 8px;
+
+.NavMobile {
+  background: linear-gradient(135deg, #ff77b4 20%, #9b4d97 50%, #ff77b4);
+  width: 100%;
+  height: 70px;
+  color: #f9f9f9;
+  display: flex;
+  align-items: center;
+  padding: 1em 2em;
+  box-sizing: border-box;
+  justify-content: center;
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  z-index: 3;
+  & ul {
+    @media screen and (max-width: 500px) {
+      justify-content: space-between !important;
+    }
+    display: flex;
+    align-items: center;
+    justify-content: space-evenly;
+    width: 100%;
+    & * {
+      cursor: pointer;
+      font-size: 2.2em;
+      transition: all 0.3s;
+      &:hover {
+        transform: translateY(-2px);
+      }
+      &:active {
+        transform: scale(1.05);
+      }
+    }
+    & img {
+      border-radius: 50%;
+      width: 50px;
+      height: 50px;
+      object-fit: cover;
+    }
+  }
+}
+
+.NavMobile__header {
+  background: linear-gradient(135deg, #ff77b4 20%, #9b4d97 50%, #ff77b4);
+  width: 100%;
+  height: 70px;
+  color: #f9f9f9;
+  display: flex;
+  align-items: center;
+  padding: 1em 2em;
+  box-sizing: border-box;
+  justify-content: center;
+  position: absolute;
+  font-family: "Raleway";
+}
